Fix gpio submodule, prune gitignore
Over-aggressive .gitignore strikes again. I believe that because gpio/
was an entry in the .gitgnore it ended up ignoring the path that's
usually supposed to happen for a submodule and so that left the
submodule move in a not really working state. I've fixed up the
submodule here and trimmed the .gitignore to only ignore things that
this project actually produces.