Add Tests and pyproject.toml, use src/ module layout for project (#14)
* Add: Tests and pyproject.toml
* Task: Refactor project to use src layout

In an effort to begin modernizing the project for packaging and
distributing on PyPi, pyproject.toml is introduced to replace
requirements.txt, the project structure now utilizes the src/ layout,
and pytest and coverage are added and configured.

* Fix gpio submodule, prune gitignore

Over-aggressive .gitignore strikes again. I believe that because gpio/
was an entry in the .gitgnore it ended up ignoring the path that's
usually supposed to happen for a submodule and so that left the
submodule move in a not really working state. I've fixed up the
submodule here and trimmed the .gitignore to only ignore things that
this project actually produces.

---------

Co-authored-by: Theo Hill <Theo0x48@gmail.com>