Use subclasses for device specific behavior
Both Amplifier and Accessory were base classes but had special
cases for one or more of their subclasses. If we're using inheritance we
might as well use it to simplify the base class. While I was there I
reorganized things a little bit to depend less on the stationd module.