Move command socket response to end of command_handler
This pulls actually sending the responses up out of all the devices and
into a single spot. With that out of the way a number of pieces fell
away. Most device actions no longer needed a Command, and Command itself
was reduced to just the list of command strings, where those could be
just passed around instead. Specialized response functions could mostly
all be removed and turned into strings on site.