Replace GPIO submodule with gpiod pip package (#18)
* Implement devcontainer and install gpiod

gpiod relies on a linux-only system package called libgpiod. To
enable cross-platform development (macOS, etc), a containerized
development environment is needed. This introduces the dev container
environment as well as the gpiod project dependency.

* Replace GPIO submodule with gpiod

A new LineOut class is defined to lightly wrap gpiod and replaces all instances
where the deprecated GPIO submodule was used. Config.ini examples were
also updated to to reflect the need to specify the gpio chip as well as the pin.